requirements.txt
torch==2.0.1
torchvision==0.15.2
transformers==4.35.0
Pillow==10.0.0
numpy==1.26.0
pandas==2.1.0
opencv-python==4.8.0
tqdm==4.66.1
scikit-learn==1.3.0
tiffslide==0.5.0